Tests extended and errors repaired in: - ConstantFloat (loaded an int instead of a float on the stack... ahem...) - GOTOW (renamed into GOTO_W as it should be called) - ICONST_m1 (renamed into ICONST_M1 as it should be called) - IF_ICMPNE (had a small copy/paste error) - INVOKEDYNAMIC (was missing, and is right now a copy of INVOKESTATIC -- will be repaired later) - InterpreterFactory (to include INVOKEDYNAMIC)

+++ trunk/dsol-base/src/main/java/nl/tudelft/simulation/dsol/interpreter/classfile/ConstantFloat.java
@@ -24,7 +24,7 @@ public final class ConstantFloat extends Constant
      */
     public ConstantFloat(final Constant[] constantPool, final DataInput inputStream) throws IOException
     {
-        this(constantPool, inputStream.readInt());
+        this(constantPool, inputStream.readFloat());
     }
 
     /**
